pyproject.toml
[tool.poetry]
name = "nonebot_plugin_dorodoro"
version = "1.5.1"
description = "一个基于文字冒险的游戏插件"
authors = ["ATTomatoo"]
license = "AGPL3.0"
readme = "README.md"
homepage = "https://github.com/ATTomatoo/dorodoro"
repository = "https://github.com/ATTomatoo/dorodoro"
keywords = ["nonebot", "nonebot2", "nonebot_plugin", "dorodoro"]

packages = [{ include = "nonebot_plugin_dorodoro" }]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.3.3"
nonebot-adapter-onebot = "^2.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
